pyproject.toml
[project]
name = "wyoming-azure-keyword"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-cognitiveservices-speech>=1.46.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "wyoming>=1.7.2",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
    "setuptools>=80.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"